Reports page now detects Ansible vs other kind of reports
Before this commit, all reports would try to be shown as Ansible
reports, therefore failing. Puppet reports for instance cannot be parsed
with the same helpers as Ansible reports.

This ensures the right partial is rendered when an Ansible report is
detected (by a naive mechanism) otherwise it renders the default config
report partial.